Move SocialLogin & socialRegisterUser to AuthService
